Implemented JSON support to Script Editor
- Implemented JSON syntax (similar to Groovy - therefore a parent class was created)
- Implemented JSON highlighting
- Implemented JSON auto-completor (dummy since there is no auto-complete)
- Improved javadocs
