+++ qupath-extension-script-editor/src/main/resources/scripting_styles.css
@@ -35,3 +35,6 @@
 .semicolon {
     -fx-font-weight: bold;
 }
+.remaining {
+    -fx-fill: darkorange;
+}
